:truck: Move cli parameters to owm module
